feat: make galax-to-gala conversion public
