+++ .pre-commit-config.yaml
@@ -33,7 +33,6 @@ repos:
       - id: mixed-line-ending
       - id: name-tests-test
         args: ["--pytest-test-first"]
-        exclude: '^tests\/.*_helper\.py$'
       - id: trailing-whitespace
 
   - repo: https://github.com/pre-commit/pygrep-hooks
